requirements.txt
requests==2.31.0
pandas==2.2.2
pillow==10.3.0
isodate==0.6.1